+++ Signum.Engine/CodeGeneration/WindowsCodeGenerator.cs
@@ -37,6 +37,8 @@ namespace Signum.Engine.CodeGeneration
             if (!Directory.Exists(projectFolder))
                 throw new InvalidOperationException("{0} not found. Override GetProjectFolder".Formato(projectFolder));
 
+            bool? overwriteFiles = null;
+
             foreach (var mod in GetModules())
             {
                 string str = WriteClientFile(mod);
@@ -44,7 +46,9 @@ namespace Signum.Engine.CodeGeneration
                 {
                     string fullFileName = Path.Combine(projectFolder, GetClientFileName(mod));
                     FileTools.CreateParentDirectory(fullFileName);
-                    File.WriteAllText(fullFileName, str);
+
+                    if (!File.Exists(fullFileName) || SafeConsole.Ask(ref overwriteFiles, "Overwrite {0}?".Formato(fullFileName)))
+                        File.WriteAllText(fullFileName, str);
                 }
 
                 foreach (var t in mod.Types)
@@ -54,9 +58,15 @@ namespace Signum.Engine.CodeGeneration
                     {
                         string fullFileName = Path.Combine(projectFolder, GetViewFileName(mod, t));
                         FileTools.CreateParentDirectory(fullFileName);
-                        File.WriteAllText(fullFileName, view);
+                        if (!File.Exists(fullFileName) || SafeConsole.Ask(ref overwriteFiles, "Overwrite {0}?".Formato(fullFileName)))
+                            File.WriteAllText(fullFileName, view);
+
+
+                        fullFileName += ".cs";
+
                         string codeBehind = WriteViewCodeBehindFile(t);
-                        File.WriteAllText(fullFileName + ".cs", codeBehind);
+                        if (!File.Exists(fullFileName) || SafeConsole.Ask(ref overwriteFiles, "Overwrite {0}?".Formato(fullFileName)))
+                            File.WriteAllText(fullFileName + ".cs", codeBehind);
                     }
                 }
             }
